genjs.ml -> genlua.ml : Down the Rabbit Hole
Recently, I've become somewhat obsessed with using Haxe to target Lua.
Partially, I want to be able to use Lua as a scripting language for Vim
and maybe NeoVim.  There exists projects like the javascript-generator
based Luaxe, but part of me wondered "how hard would it be to just do a
proper lua target for Haxe?".  This branch will serve to scratch that
itch, and I'll document my progress along the way.

The javascript generator already comes pretty close to handling all of
the cases I need, it's just a little schizophrenic.  Sometimes certain
code won't work because it thinks it's still in the js namespace, etc.

So, since javascript comes so close to lua, we'll use it as a base.
The first step is to copy the existing genjs.ml over to genlua.ml.  From
there, I can track the changes to genlua.ml over the existing javascript
generator.