Haxe, meet Lua
The next step is to let the Haxe compiler "know" about the new lua
target available via genlua.ml.  This involves adding genlua to the
Makefile, and introducing Lua as an enum for the purposes of various
command and platform behaviors.

At ths point, you can invoke "haxe -lua" just fine, but the resulting
code will complain about a lack of an std namespace for certain types.

We'll need to do something for std like we did for genlua... copy
existing js files over until the compiler is happy.  That's coming up
next.