Getting rid of irrelevant features : Source mapping, js modern/flatten
Javascript is one of the more flexible targets in the Haxe language.
There's actually two different ways you can generate javascript source.
(Without going into too many details, one is for legacy purposes).  In
addition, there's also the ability to flatten packages, and to generate
source maps for certain browser debugging purposes.

None of those features are appropriate for Lua, and they constituted a
fair amount of code in the original genjs source. So, it's a good idea
to get rid of them early and focus on what matters.