Lua needs a bit of help
For some reason, the creators of Lua decided not to add bitwise
operations to their library until version 5.3.  There's still a lot of
installations out there that will have versions below that, and in those
cases the Lua creators suggest to use the "bit" library.

So, it's necessary to do some more operator hacks to get lua to play
nicely with simple expressions like ( 4 | 5);

This gets a little more complicated due to the fact that many bit
operators can also be assignment operators (|=, etc.)  So, in order to
keep the code clean it's necessary to create a few helper functions.

As a result, the lua generator can transform operators that look like
this:

```haxe
var x = (4 | 5);
```

into this:

```lua
local x = bit.or(4,5);
```