Express your Statements
Most pieces of code in Haxe are actually expressions.  This means that
certain operations can be combined, such as chained assignments:

```haxe3
var x = 0;
var y = 0;

x = y = 3;
```

Lua define certain operations as statements, meaning that they don't
return values, and can't be combined with other expressions.

The best way that I can see to deal with this is to wrap lua statements
inside of immediately invoked function expressions that execute the
statement logic, and then return the result.

`` lua
    x = (function() y = 3 return y end)()
```

To do this, we need to reach into the variable declaration logic, and
have it reuse the iife logic for tbinops.