Working on Lua Boot.hx : Trace all the things!
One of the next things I wanted to be able to do was to have a decent
way of tracing arbitrary variables like arrays and objects.  Lua
actually doesn't make a distinction between arrays and objects, which
means I'll need to doctor up a lua array with some extra metadata in
order for reflection-based tracing to work.  That'll be for another time
though.