Lua: Blocked by blocks, take two.
My previous commit tried to provide Haxe blocks in a Lua context.
Earlier, I said that an IIFE is the right way to provide a block-like
feature in Lua, since it doesn't support them.

It turns out, I was wrong in three ways.  Blocks *are* provided by Lua
with simple do/end syntax:

```lua
do
    local some_var = 4
end;
```

Iifes are actually a bad way of providing block functionality, because
a "return" command is valid inside of a block, which should return from
the parent function context.  If the block is an iffe function itself,
it won't have the right behavior.

Finally, it would've been fine just to stick with javascripts method of
creating "virtual" blocks by aliasing variable assignments.  That works
for every case I can come up with.  The only problem is that in some
cases the javascript generator will produce a bare value.  E.g.:

```lua
   var x = 4;
   4;
```

This is an artifact of some of the getter/setter and abstract behavior.
In javascript, the "4" by itself is harmless.  But, in Lua, it's not
allowed.

So, I'm cleaning this up by going back to the "virtual" block approach,
and ignoring no-op variable expressions.  I might end up using lua
blocks eventually if I see that they provide some other benefit.  But,
since they declare a new scope, they actually slow things down a bit.