Lua constructors
Lua isn't object oriented by default, but luckily it has a "colon"
method syntax that allows for easier instance method specification.  Lua
uses "self" instead of "this", so relevant keyword usage needs to be
updated.  Finally, it doesn't really have a constructor, so the "new"
method will need to build and return a simple table.