Multiple problems with multiple inheritance
Lua has a neat "metatable" implementation that lets you provide a
fallback lookup for fields that don't directly exist on a given table
(object).  This is pretty close to what we need for inheritance,
primarily inheritance of methods, etc.  However, this metatable lookup
is not recursive (like js prototypes), making multiple inheritance problematic.

We don't have the same problem for instance fields set by the
constructor, we just can set the new instance "self" to the value
returned by the parent constructor.  All of the fields in the parents
will then be set appropriately.

For methods, I've set up an empty __methods table on the instance pertaining to
a given class. I've also used the code for the old js "prototype" object
to define an "mt" table that defines all of the instance methods.  I'll
have to rename it, as mt is probably too common a word to claim.

The mt table contains the functions (for example):

```lua
Foo.mt = {
    someFunc  = function() end
}
Bar.mt = {
    someOtherFunc = function() end
}
```

Anyways, back in the constructor, the "self" object will add its current
class "mt" object to its __methods table.  This also happens for each of
its parent class instances.  For instance, the Foo instance would make a
call like this:

```lua
 table.insert(self.__methods, Foo.mt);

```

Finally, the "self" instance in the constructor is assigned a metatable
with a defined index:

```lua
setmetatable(self, lua.Boot.resolveMethod)
```

This method defined in Boot will be passed the current table instance.
It will proceed to look through each of the __methods tables in turn and
try to find the requested field name (key).
It will look like:

```lua
lua.Boot.resolveMethod = function(table, key)
    for index, value in ipairs(table.__methods) do
        if value[key] ~= nil then return value[key] end
    end
    return nil
end
``

Of course, if resolveMethod finds the field, it will return it, allowing
the method to be called.  Otherwise, it just will return nil.  That
should never happen though, since the field is checked at compile time.