Lua: TableTools and LuaStringTools
I made some quick externs of some of the basic lua string/table
manipulation routines.  I need to use these instead of untyped table.<x>
because lower case names like that are inferred to be instance variables
by Haxe, and instance methods get the "table:<x>" colon treatment.

This means that I need to extern pretty much all of the Lua helper
classes, but that's not such a bad thing.