Lua : Mapping out Maps
Haxe has a complex abstract type for dealing with standard Map<A,B>
types.  This is because many targets require completely different
datastructures for dealing with int/string based maps.

Luckily, here's where things are actually easier with Lua : Lua only has
one type of map.  So, we can use that for both IntMap and StringMap.

There's a few other gotchas: Lua can't tell if a certain key is set or
not.  It's necessary to maintain a separate table of key lookups there.

I also fixed an abstract related metadata field __interfaces__ in
genlua, so that it would use a normal anonymous table.  This doesn't
need to be an actual haxe array, since the entries will never be null,
and it won't be directly exposed by a Haxe language feature.