Lua : Add support for throwing errors.
Lua uses an "error" method to throw an exception, instead of "throw".
So, we just need to change the word a bit, and call it as a function.

Note that we're not handling these exceptions with try/catch.  Lua
doesn't have those either, and they'll require additional work to get
something equivalent working.