Lua: Map redux
It looks like typer.ml expects Map to be an abstract in some cases.
I'll need to make some adjustments so that everything works.
Essentially, I'm moving the base Map implementation to the lua
namespace.  StringMap, IntMap, etc. are typedefs for this impl.

The std Map will be a standard abstract over those types.