Lua: Back to the drawing board with strings.
It was clear after running some tests that simply inlining the
implementation for strings is not going to cut it for the tests.  There
are situations involving type unification for structural types that
require an accessible field.  So, I've taken a different tack for
String.  This time, I'm adding the haxe method metatable to the base
string metatable.  Luckily, there's no conflicts.  The primary lua string
methods will still be reached, and then anything not in there will be
looked up against the haxe implementation. Modifications in Boot and
String take care of this.  There's another tweak to Std that I needed to
make, for some reason the cast type needed to be made explicit.