Lua: trying to catch more edge cases
Lua doesn't have a conventional try/catch mechanism, but it does have
something that comes pretty close: pcall.

The pcall method works by passing it a function argument, which it
immediately executes.  The pcall method returns two values:  The first
value is the function execution status.. true means the function
completed successfully, false means an error was thrown.  The second
value returned by pcall is the successful return, or the thrown error
(respectively).

There's a few modifications necessary for interop with normal Haxe
syntax here.  First of all, Haxe "try" syntax can work on arbitrary
blocks of code, not necessarily just functions.  The lua generator needs
to create a small iife to execute the code as if it were a block.  It
then can catch errors normally.

However, that's not the end of the story... the "try" block in Haxe
itself may itself contain a return.  Normally, an iife doesn't have a
return value.  But, in this case, we'll have it return an "expected
return value".  If the function returns succesfully, and the return
value differs from what's expected, we'll return that new value instead.

Note that I'm using a lot of statically assigned temporary variables
here.  This isn't best practice, but I made a TODO to come back and
clean it up with tmp variable names guaranteed not to conflict.