Lua: Workaround for invoking instance methods on constants
In Haxe, you can invoke a method directly on an anonymous value type,
such as "foo bar".charAt(0).  However, in Lua, this isn't possible.
Once again, I'm wrapping the expression in an iife.

I've been using iife's a lot to work around missing features of ua.
This is somewhat problematic because closures like this are slow in Lua.
in this case it might be a better idea to try and inline the basic
String functions.  That caused some problems with some of the tests, but
there may still be a way forward there.