Lua : OOP tweaks
Previously, I was using an array to store the inherited method
metatables for a given instance.  It's actually a lot cleaner just to
use __index recursively on tables.  I've made the necessary changes.

I've also cleaned up some of the metatable generation code, and avoid
creating the method index table instance if there are no methods for the
current class.