Lua : iife wrap a few generated block elements
In a few cases, Haxe will generate simple field accesses as a complete
statement:

```haxe
var a = { f : 4}
a.f;
```

In Lua, you can't have a no-op statement like this.  However, there's a
bit of uncertainty about whether or not this is truly no-op... e.g. it
might actually trigger a field getter, etc.  So, to be truly safe we
must wrap this in an iife.

I'm refactoring some of the created iife methods into a helper.  It
doesn't cover many of them yet, there's too many variations to cover
them all for now.