Lua : More oop fixes.
One more iteration, I think I like this method the best so far.

The constructors are simplified here... it's no longer necessary to pass
in an empty table object to perform construction on.

However, rather than performing the actual construction in the "new"
method, now this is done in a autogenerated __construct method.  This
method *does* accept a self method, and performs the actual field
manipulation for the current constructor.

This enables multiple constructors to be called via "super()", while
keeping the actual public constructor in Lua more closely related to the
haxe equivalent.